% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSird.R
\name{multiEpochSird}
\alias{multiEpochSird}
\title{Multi-epoch SIRD}
\usage{
multiEpochSird(
  n.rep = 10,
  paramsList,
  timesList,
  interventionList,
  init,
  nThreads = 1,
  func = sird,
  modelOutput = if (is.list(init) & !is.data.frame(init)) names(init[[1]]) else
    names(init)
)
}
\arguments{
\item{n.rep}{The number of simulation replicates. Defaults to 10.}

\item{paramsList}{A list of parameter value lists. Each list entry must match \code{params} from \code{mcSeirScenario}.}

\item{timesList}{A list of time vectors.  Each list entry must match \code{times} from \code{mcSeirScenario} and be consecutive.}

\item{interventionList}{A list of intervention vectors. Each list entry must match \code{intervention} from \code{mcSeirScenario}.}

\item{init}{Initial values for \code{S}, \code{E}, \code{I}, \code{R} and if \code{func = seira}, \code{A}, the ambient source of infection.}

\item{nThreads}{The number of threads to use in multi-core computing.}

\item{func}{The function defining the derivatives in the ODE system (which define the model). Defaults to \code{seira}.}
}
\description{
Conducts simulations for a multi-epoch SIRD model
}
