% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSird.R
\name{seir}
\alias{seir}
\title{SEIR Model}
\usage{
seir(times, state, parameters)
}
\arguments{
\item{times}{The times at which to evaluate the solutions.}

\item{state}{The current estimate of the variables in the ODE system.}

\item{parameters}{A vector of parameters. Must include \code{beta}, \code{rho}, \code{L} and \code{gamma}.}
}
\description{
This function defines the system of ODEs for a SEIR model. For use with deSolve::ode.
}
