% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSird.R
\name{sird}
\alias{sird}
\title{SIRD Model}
\usage{
sird(times, state, parameters)
}
\arguments{
\item{times}{The times at which to evaluate solutions.}

\item{state}{The current estimate of the variables in the ODE system.}

\item{parameters}{A vector of parameters. Must include parameters \code{t_s}, \code{a_s}, \code{b_s}, and \code{c_s}.}
}
\description{
This function defines the system of ODEs for the SIRD model. For use with deSolve::ode.
}
